# swagger.yaml
definitions:
...
  ErrorResponse:
    description: ErrorResponse defines the standard error response structure
    properties:
      code:
        type: string
        x-go-name: Code
      details:
        items:
          $ref: "#/definitions/ValidationError"
        type: array
        x-go-name: Details
      error:
        type: string
        x-go-name: Error
      message:
        type: string
        x-go-name: Message
      success:
        type: boolean
        x-go-name: Success
      timestamp:
        type: string
        x-go-name: Timestamp
    type: object
    x-go-package: github.com/purnama354/sejiwa-api/internal/dto
...
  SuccessResponse:
    description: SuccessResponse defines the standard success response structure
    properties:
      message:
        type: string
        x-go-name: Message
      success:
        type: boolean
        x-go-name: Success
      timestamp:
        type: string
        x-go-name: Timestamp
    type: object
    x-go-package: github.com/purnama354/sejiwa-api/internal/dto
...
  ValidationError:
    description: ValidationError defines individual field validation errors
    properties:
      code:
        type: string
        x-go-name: Code
      field:
        type: string
        x-go-name: Field
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/purnama354/sejiwa-api/internal/dto
info:
  title: Sejiwa API
  description: Mental health discussion platform API
  version: "1.0.0"
paths:
...
  /moderation/threads/{id}:
    post:
      description: Moderate a thread (Moderator/Admin)
      operationId: moderateThread
      parameters:
        - description: Thread ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
      responses:
        "200":
          $ref: "#/responses/successResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "422":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - moderation
...
responses:
...
  errorResponse:
    description: Generic error envelope
    schema:
      $ref: "#/definitions/ErrorResponse"
...
  successResponse:
    description: Generic success envelope
    schema:
      $ref: "#/definitions/SuccessResponse"
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    in: header
    name: Authorization
    description: "JWT token in the format 'Bearer {token}'"
swagger: "2.0"